requirements.txt
# core
coverage (3.6)
Django (1.7)
django-allauth (0.19.1)
oauthlib (0.7.2)
Pillow (2.7.0)pip 
(6.0.8)
python-openid (2.2.5)
requests (2.6.0)
requests-oauthlib (0.4.2)
setuptools (12.0.5

# testing
#mock
#factory-boy>=1.3,<2.0

# docs
#Sphinx
